Move elastic search logic to express app.
