+++ express/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "export NODE_ENV=development && nodemon ./bin/www | bunyan -o short",
     "prod": "export NODE_ENV=production && forever start ./bin/www",
-    "prod-fg": "export NODE_ENV=production && forever ./bin/www",
+    "prod-fg": "export NODE_ENV=production && forever ./bin/www | bunyan -o short",
     "lint": "eslint *.js **/*.js; exit 0",
     "lint-html": "eslint *.js **/*.js -f html -o ./reports/lint-results.html; exit 0",
     "lint-fix": "eslint --fix *.js **/*.js; exit 0"
@@ -16,6 +16,7 @@
     "bunyan": "^1.8.5",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.0",
+    "elasticsearch": "^12.1.3",
     "express": "~4.14.1",
     "jade": "~1.11.0",
     "morgan": "~1.7.0",
